backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.2
djangorestframework==3.16.0
drf-yasg==1.21.10
h11==0.16.0
idna==3.10
inflection==0.5.1
mysqlclient==2.2.7
packaging==25.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
